refactor(entry points): * merged back different index-xx into a single entry point * removed all redundant code * moved a few more utility functions to helpers * cleanups & minor fixes
